+++ tests/APITest.php
@@ -167,6 +167,597 @@ class APITest extends BaseTest {
     }
 
     /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRawRequest() {
+        $request = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isValidResponse($response, $request);
+        $this->assertInternalType('array', $response['result']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithAdditionalHeaders() {
+        $request = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $headers = [
+            'X-Test-Header' => 'ABC123'
+        ];
+
+        $response = $this->api->rawRequest($request, $headers);
+
+        $requestHeaders = explode("\r\n", $this->api->getLastRequestHeaders());
+        $this->assertContains('X-Test-Header: ABC123', $requestHeaders);
+
+        // Additional checks:
+        $this->assertInternalType('array', $response);
+        $this->isValidResponse($response, $request);
+        $this->assertInternalType('array', $response['result']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testInvalidRequests() {
+        $invalidRequests = [
+            [],
+            [1],
+            [1, 2, 3]
+            // @todo Unable to check, because rawRequest() expects an array:
+//            null,
+//            '',
+//            new \stdClass(),
+//            true,
+//            false,
+//            $this->generateRandomString(),
+//            23,
+//            -42,
+//            123.456
+        ];
+
+        foreach ($invalidRequests as $invalidRequest) {
+            $response = $this->api->rawRequest($invalidRequest);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->assertNull($response['id']);
+            $this->assertSame(-32600, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithMissingParameters() {
+        $request = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'id' => 1
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isError($response);
+        $this->hasValidJSONRPCIdentifier($request, $response);
+        $this->assertSame(-32602, $response['error']['code']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithInvalidParameters() {
+        $requestTpl = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'id' => 1
+        ];
+
+        $invalidParameters = [
+            null,
+            23,
+            -42,
+            123.456,
+            true,
+            false,
+            '',
+            $this->generateRandomString()
+        ];
+
+        foreach ($invalidParameters as $invalidParameter) {
+            $request = $requestTpl;
+            $request['params'] = $invalidParameter;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->hasValidJSONRPCIdentifier($request, $response);
+            $this->assertSame(-32602, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithMissingApiKey() {
+        $request = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [],
+            'id' => 1
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isError($response);
+        $this->hasValidJSONRPCIdentifier($request, $response);
+        $this->assertSame(-32099, $response['error']['code']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithInvalidApiKeys() {
+        $requestTpl = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => $this->generateRandomString()
+            ],
+            'id' => 1
+        ];
+
+        $invalidAPIKeys = [
+            null,
+            new \StdClass(),
+            [],
+            23,
+            -42,
+            123.456,
+            true,
+            false,
+            '',
+            $this->generateRandomString()
+        ];
+
+        foreach ($invalidAPIKeys as $invalidAPIKey) {
+            $request = $requestTpl;
+            $request['params']['apikey'] = $invalidAPIKey;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->hasValidJSONRPCIdentifier($request, $response);
+            $this->assertSame(-32099, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithMissingVersionNumber() {
+        $request = [
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isError($response);
+        $this->hasValidJSONRPCIdentifier($request, $response);
+        $this->assertSame(-32603, $response['error']['code']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithInvalidVersionNumbers() {
+        $requestTpl = [
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $invalidVersionNumbers = [
+            null,
+            new \StdClass(),
+            [],
+            23,
+            -42,
+            123.456,
+            true,
+            false,
+            '',
+            $this->generateRandomString()
+        ];
+
+        foreach ($invalidVersionNumbers as $invalidVersionNumber) {
+            $request = $requestTpl;
+            $request['jsonrpc'] = $invalidVersionNumber;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->hasValidJSONRPCIdentifier($request, $response);
+            $this->assertSame(-32600, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithValidIdentifiers() {
+        $requestTpl = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ]
+        ];
+
+        $validIdentifiers = [
+            23,
+            -42,
+            0,
+            $this->generateRandomString()
+        ];
+
+        foreach ($validIdentifiers as $validIdentifier) {
+            $request = $requestTpl;
+            $request['id'] = $validIdentifier;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isValidResponse($response, $request);
+            $this->hasValidJSONRPCIdentifier($request, $response);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithInvalidIdentifiers() {
+        $requestTpl = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ]
+        ];
+
+        $invalidIdentifiers = [
+            null,
+            new \StdClass(),
+            [],
+            123.456,
+            true,
+            false,
+            ''
+        ];
+
+        foreach ($invalidIdentifiers as $invalidIdentifier) {
+            $request = $requestTpl;
+            $request['id'] = $invalidIdentifier;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->assertNull($response['id']);
+            $this->assertSame(-32600, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithMissingMethod() {
+        $request = [
+            'jsonrpc' => '2.0',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isError($response);
+        $this->hasValidJSONRPCIdentifier($request, $response);
+        $this->assertSame(-32601, $response['error']['code']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithInvalidMethods() {
+        $requestTpl = [
+            'jsonrpc' => '2.0',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $invalidMethods = [
+            null,
+            new \StdClass(),
+            [],
+            23,
+            -42,
+            123.456,
+            true,
+            false
+        ];
+
+        foreach ($invalidMethods as $invalidMethod) {
+            $request = $requestTpl;
+            $request['method'] = $invalidMethod;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->hasValidJSONRPCIdentifier($request, $response);
+            $this->assertSame(-32600, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRequestWithUnkownMethods() {
+        $requestTpl = [
+            'jsonrpc' => '2.0',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ],
+            'id' => 1
+        ];
+
+        $unknownMethods = [
+            $this->generateRandomString(),
+            'cmdb.nope',
+            ''
+            // @todo i-doit's routing thinks it's "cmdb.objects.read" (even if "read" doesn't exist):
+            //'cmdb.objects.' . $this->generateRandomString()
+        ];
+
+        foreach ($unknownMethods as $unknownMethod) {
+            $request = $requestTpl;
+            $request['method'] = $unknownMethod;
+
+            $response = $this->api->rawRequest($request);
+
+            $this->assertInternalType('array', $response);
+            $this->isError($response);
+            $this->hasValidJSONRPCIdentifier($request, $response);
+            $this->assertSame(-32601, $response['error']['code']);
+        }
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testRepeatingIdentifiersInBatchRequest() {
+        $request = [
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ],
+                'id' => 1
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ],
+                'id' => $this->generateRandomString()
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ],
+                'id' => 1
+            ]
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isError($response);
+        $this->assertNull($response['id']);
+        $this->assertSame(-32603, $response['error']['code']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testVariousApiKeysInBatchRequest() {
+        $request = [
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ],
+                'id' => 1
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => $this->generateRandomString()
+                ],
+                'id' => 2
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => $this->generateRandomString()
+                ],
+                'id' => 3
+            ]
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->isError($response);
+        $this->hasValidJSONRPCIdentifier($request, $response);
+        $this->assertSame(-32602, $response['error']['code']);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testNotification() {
+        $request = [
+            'jsonrpc' => '2.0',
+            'method' => 'idoit.version',
+            'params' => [
+                'apikey' => getenv('KEY')
+            ]
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertEmpty($response);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testOnlyNotificationsInBatchRequest() {
+        $request = [
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ]
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'cmdb.object.read',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ]
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'cmdb.category.read',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ]
+            ]
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertEmpty($response);
+    }
+
+    /**
+     * @group unreleased
+     * @throws \Exception on error
+     */
+    public function testSomeNotificationsInBatchRequest() {
+        $request = [
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'idoit.version',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ],
+                'id' => 1
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'cmdb.object.read',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ]
+            ],
+            [
+                'jsonrpc' => '2.0',
+                'method' => 'cmdb.category.read',
+                'params' => [
+                    'apikey' => getenv('KEY')
+                ],
+                'id' => 2
+            ]
+        ];
+
+        $response = $this->api->rawRequest($request);
+
+        $this->assertInternalType('array', $response);
+        $this->assertCount(2, $response);
+
+        $this->assertArrayHasKey(0, $response);
+        $this->isValidResponse($response[0], $response[0]);
+
+        $this->assertArrayHasKey(1, $response);
+        $this->isError($response[1]);
+        $this->hasValidJSONRPCIdentifier($request[2], $response[1]);
+    }
+
+    /**
      * @covers ::getLastInfo
      * @throws \Exception
      */