+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 -   `CheckMKTags`: Read host tags by one or more objects from category `C__CATG__CMK_TAG`
+-   `API::rawRequest()`: Perform a low level API request
 
 
 ### Changed